==> 2_Combined-data-processing.ipynb <==
import pandas as pd
import numpy as np
import os
import glob
from pprint import pprint
from itertools import islice
import random

# Set the max_columns option to None to display all columns
pd.set_option('display.max_columns', None)

# List of state abbreviations
state_abbreviations = ['AZ', 'CA', 'CO', 'ID', 'MT', 'NM', 'NV', 'OR', 'UT', 'WA', 'WY']

bea_df = pd.read_csv('/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/BEA_economics_CAEMP25N_clean/BEA_combined_2001_2021.csv')
census_df = pd.read_csv('/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/census_clean/census_2001_2021_clean.csv')

print(len(bea_df))
print(len(census_df))

merge_df = pd.concat([census_df, bea_df], ignore_index=True)

len(merge_df)*2

### Calculate the normalized jobs_per_100k people values per year

jobsper100k_new_rows = []  # Step 1: Create an empty list for new rows

# Set the maximum number of iterations you want
num_iterations = len(merge_df['geofips'].unique())  # Adjust this to your desired number of iterations

# Create an iterator over the unique 'geofips'
geofips_iterator = iter(merge_df['geofips'].unique())

# Step 2: Iterate through unique 'geofips'
for geofips in islice(geofips_iterator, num_iterations):
    # Extract data for the current 'geofips'
    subset = merge_df[merge_df['geofips'] == geofips]
#     print(f"Temporary df for geofips {geofips}:")
#     print("***///***")

    # Create a new row with the calculated values
    new_row = subset.iloc[0].copy() # copies one of the existing rows for this geofips

    new_row['description'] = 'jobs_per_100k' # updates the description column
    new_row['linecode'] = 30.0  # updates the linecode    

    # Calculate 'jobs_per_100k' for each year
    for year in range(2001, 2022):
        column = f'yr_{year}'
        linecode_jobs = 10.0
        linecode_population = 20.0


        # Calculate 'jobs_per_100k' for this year
        jobs_data = subset[subset['linecode'] == linecode_jobs][column].values[0]
        population_data = subset[subset['linecode'] == linecode_population][column].values[0]

        jobs_per_100k = round((jobs_data / population_data) * 100000)

#         print(column)
#         print(jobs_data)
#         print(population_data)
#         print(jobs_per_100k)            
#         print("***///***")


        new_row[column] = jobs_per_100k  # Set the jobs_per_100k value for current year

#         print(new_row)
#         print("***///***")        
        
    # Append the new row to the list after all years have been updated with the jobs_per_100k values
    jobsper100k_new_rows.append(new_row)

merge_df[merge_df["geofips"]==4001]

# Convert the list of new rows into a DataFrame
new_data = pd.DataFrame(jobsper100k_new_rows)

new_data

# Concatenate the new data with the original "merge_df" using pd.concat
merge_df_two = pd.concat([merge_df, new_data], ignore_index=True)

# If you want to sort the resulting DataFrame by a specific column (e.g., 'geofips' and 'year'):
merge_df_two = merge_df_two.sort_values(by=['geofips', 'description'])

# Reset the index for the updated DataFrame
merge_df_two.reset_index(drop=True, inplace=True)

merge_df_two[merge_df_two["geofips"]==4001]

### Calculate the annual jobs growth rate for each county from 2002-2021
annual_growth_rate = (jobs_per_100k_yearY - jobs_per_100k_yearX) / jobs_per_100k_yearY

ann_growthrate_new_rows = []

# Iterate through unique geofips
for geofips in merge_df_two['geofips'].unique():
    # Extract data for the current geofips
    subset = merge_df_two[merge_df_two['geofips'] == geofips]

    # Create a new row with the calculated values
    new_row = subset.iloc[0].copy()
    new_row['description'] = 'ann_growthrate'
    new_row['linecode'] = 40.0    
    
    tot_growthrate_5y = 0  # Initialize the total annual growth rate for 5 years
    tot_growthrate_10y = 0  # Initialize the total annual growth rate for 10 years
    tot_growthrate_20y = 0  # Initialize the total annual growth rate for 20 years
    
    # Iterate through years to calculate annual growth rates
    for year in range(2001, 2022):
        past_column = f'yr_{year-1}'
        current_column = f'yr_{year}'

        if year == 2001:
            # Set ann_growthrate value for yr_2001 to "NA"
            new_row[current_column] = "NA"
        else:        
            # Calculate annual growth rate for the current year
            jobs_per_100k_past = subset[past_column].values[0]
            jobs_per_100k_current = subset[current_column].values[0]

            ann_growthrate = (jobs_per_100k_current - jobs_per_100k_past) / jobs_per_100k_current
            new_row[current_column] = ann_growthrate

            # Accumulate the annual growth rates
            tot_growthrate_20y += ann_growthrate
            
            if 2017 <= year <= 2021:
                tot_growthrate_5y += ann_growthrate
            if 2012 <= year <= 2021:
                tot_growthrate_10y += ann_growthrate

    # Calculate the average annual growth rates for 5 years, 10 years, and 20 years
    avg_growthrate_5y = tot_growthrate_5y / len(range(2017, 2022))
    avg_growthrate_10y = tot_growthrate_10y / len(range(2012, 2022))            

    # Calculate the average annual growth rate
    avg_growthrate_20y = tot_growthrate_20y / len(range(2002, 2022))  # Since growthrate data begins in 2002

    # Append the average annual growth rates to the new row
    new_row['avg_growthrate_5y'] = avg_growthrate_5y
    new_row['avg_growthrate_10y'] = avg_growthrate_10y
    new_row['avg_growthrate_20y'] = avg_growthrate_20y           
            
    # Append the new row to the list
    ann_growthrate_new_rows.append(new_row)

# Convert the list of new rows into a DataFrame
new_data = pd.DataFrame(ann_growthrate_new_rows)

# Concatenate the new data with the original "merge_df" using pd.concat
final_df = pd.concat([merge_df_two, new_data], ignore_index=True)

# If you want to sort the resulting DataFrame by a specific column (e.g., 'geofips' and 'year'):
final_df = final_df.sort_values(by=['geofips', 'description'])

# Reset the index for the updated DataFrame
final_df.reset_index(drop=True, inplace=True)

final_df

# Create the output file path for the combined data
output_file_path = '/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/df_foranalysis.csv'

# Save the combined data to the output CSV file
final_df.to_csv(output_file_path, index=False)

# For testing, if necessary

# # Get a random sample of geofips identifiers
# random_seed = 427
# np.random.seed(random_seed)

# random_geofips = random.sample(final_df['geofips'].unique().tolist(), 3)

# # Create a mask to filter rows associated with the selected geofips
# mask = final_df['geofips'].isin(random_geofips)

# # Filter the DataFrame to include all rows associated with the selected geofips
# test_csv = final_df[mask].sort_values(by=['geofips', 'description'])

# # Create the output file path for the combined data
# output_file_path = '/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/test_csv.csv'

# # Save the combined data to the output CSV file
# test_csv.to_csv(output_file_path, index=False)
