==> 3_Combine_counties_economics.ipynb <==
import geopandas as gp
import pandas as pd
import numpy as np
import os
import glob
from pprint import pprint
from itertools import islice
import random

# Set the max_columns option to None to display all columns
pd.set_option('display.max_columns', None)

# set common crs for project
# epsg3310: https://epsg.io/3310-1739
# units: meters
meters_crs = 3310  # Projected crs. this should be good for this overlay() calculation and all of project. 


economic_df = pd.read_csv('/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/df_foranalysis.csv')
us_counties = gp.read_file("/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/tl_2019_us_county/tl_2019_us_county.shp")  # CA counties


us_counties['GEOID'] = us_counties['GEOID'].astype(int)

# merge
merged_df = economic_df.merge(us_counties[['GEOID', 'geometry']], left_on='geofips', right_on='GEOID', how='left')

merged_df = merged_df.drop(columns='GEOID')

gdf_merged = gp.GeoDataFrame(merged_df, geometry='geometry')

gdf_merged.head()

gdf.head(2)

gdf = gdf.rename(columns={'avg_growthrate_5y': 'avg_gr_5y'})
gdf = gdf.rename(columns={'avg_growthrate_10y': 'avg_gr_10y'})
gdf = gdf.rename(columns={'avg_growthrate_20y': 'avg_gr_20y'})

# subset to just the average growth rate data for ease of joining tables in ArcGIS
gdf_avggrowthrate = gdf[gdf['linecode'] == 40].copy()
gdf_avggrowthrate.to_file('/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/geodf_growthrate.shp', driver='ESRI Shapefile')

# gdf.to_file('/Users/natebender/Documents/ArcGIS/Projects/Western_protected_lands/geodf_foranalysis.shp', driver='ESRI Shapefile')
